requirements.txt
# Core Application Dependencies
fastapi
uvicorn[standard]  # Installs uvicorn with necessary dependencies
pydantic
requests

# Document Processing & RAG Pipeline
PyMuPDF
tabulate
python-docx
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-pinecone
llama-index-graph-stores-neo4j
pinecone-client
openai
tiktoken # Used by OpenAI models for token counting
neo4j
httpx
tenacity

# Utility & Deployment (Optional but recommended)
python-dotenv
gunicorn